Implements getMeals filter by calories and getAllMeals

+++ src/main/java/com/techreturners/teama/healthyfood/api/service/HealthyFoodService.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface HealthyFoodService {
 
-    List<Meal> getMeals(Long calories, List<String> excludedIngredients, List<String> excludedDiets, String category);
+    List<Meal> getMeals(int calories, List<String> excludedIngredients, List<String> excludedDiets, String category);
 
     List<Ingredient> getAllIngredients();
     Ingredient getIngredientById(Long id);
